fix(infra/jwt): :bug: update userId type from uint64 to uint

+++ internal/pkg/infra/jwt/service.go
@@ -135,7 +135,7 @@ func (s *service) ValidateToken(c *gin.Context) error {
 	}
 
 	c.Set("tokenId", claims["jti"])
-	c.Set("userId", userId)
+	c.Set("userId", uint(userId))
 
 	return nil
 }
